move File related code to its own modules. close #519
This creates the files file_mgr.py and file_mgr_view.py. This only moves
the files class and its helpers to the new files. It doesn't make any
changes to the code. The imports did get cleaned up a bit tho.
