+++ src/book_ease.py
@@ -23,23 +23,17 @@
 #  MA 02110-1301, USA.
 """Entry point for book_ease program"""
 import os
-from datetime import datetime
 import re
 import logging
-from pathlib import Path
-#pylint: disable=unused-import
-import pdb
-#pylint: enable=unused-import
 import gi
-#pylint: disable=wrong-import-position
-gi.require_version("Gtk", "3.0")
-#pylint: enable=wrong-import-position
+gi.require_version("Gtk", "3.0")  # pylint: disable=wrong-import-position
 from gi.repository import Gtk, GdkPixbuf, Gdk
 from gi.repository.GdkPixbuf import Pixbuf
-import signal_
 import book_reader
 import book_ease_tables
 import player
+import file_mgr
+from gui.gtk import file_mgr_view
 
 logging_stream_handler = logging.StreamHandler()
 logging_stream_handler.setFormatter(logging.Formatter('%(levelname)s - %(name)s - %(message)s'))
@@ -449,331 +443,6 @@ class Image_View:
         return(dest_width, dest_height)
 
 
-class FilesDBI:
-    """Adapter to help Files interface with book_ease.db"""
-
-    def __init__(self):
-        self.settings_string = book_ease_tables.SettingsString()
-
-    def get_library_path(self) -> str | None:
-        """get the saved path to the root directory of the book library"""
-        with book_ease_tables.DB_CONNECTION_MANAGER.query() as con:
-            library_path_list = self.settings_string.get(con, 'Files', 'library_path')
-        return library_path_list[0]['value'] if library_path_list else None
-
-    def set_library_path(self, library_path: str):
-        """
-        Save the path to the root directory of the book library
-        Not yet Implemented.
-        """
-        pass
-
-
-class Files_(signal_.Signal):
-    """class to manage the file management features of book_ease"""
-    default_library_path = str(Path.home())
-
-    def __init__(self):
-        signal_.Signal.__init__(self)
-        files_dbi = FilesDBI()
-        self.current_path = files_dbi.get_library_path() or self.default_library_path
-        self.path_back_max_len = 10
-        self.path_ahead_max_len = 10
-        self.path_back = []
-        self.path_ahead = []
-        self.show_hidden_files = False
-        self.sort_ignore_case = True
-        self.sort_dir_first = True
-
-        self.file_list = self.get_file_list_new()
-        self.icon_pos, self.f_name_pos, self.is_dir_pos, self.f_size_pos, self.f_units_pos, self.ctime_pos \
-            = (0, 1, 2, 3, 4, 5)
-
-        # Signals
-        # Notify of file changes
-        self.add_signal('cwd_changed')
-        # populate the file_list
-        self.__update_file_list()
-
-    def __update_file_list(self):
-        """repopulate the files list gtk model with the files in cwd"""
-        self.file_list.clear()
-        self.populate_file_list(self.file_list, self.current_path)
-        # notify subscribers that the file list has been updated
-        self.send('cwd_changed')
-
-    def get_file_list_new(self):
-        """create a new file list model for the files view"""
-        f_list = Gtk.ListStore(Pixbuf, str, bool, str, str, str)
-        return f_list
-
-    def populate_file_list(self, file_list, path):
-        """Determine if files in path, directory, are suitable to be displayed and add them to the file_list"""
-        files = os.scandir(path)
-        # populate liststore
-        for i in files:
-            # ignore things like broken symlinks
-            if not i.is_file() and not i.is_dir():
-                continue
-            # user option
-            if not self.show_hidden_files and self.is_hidden_file(i.name):
-                continue
-            # format timestamp
-            timestamp_formatted = datetime.fromtimestamp(i.stat().st_ctime).strftime("%y/%m/%d  %H:%M")
-            # format file size and select correct units
-            size_f, units = self.format_f_size(i.stat().st_size)
-            # set correct icon
-            icon = Gtk.IconTheme.get_default().load_icon('multimedia-player', 24, 0)
-            if i.is_dir():
-                icon = Gtk.IconTheme.get_default().load_icon('folder', 24, 0)
-            # append to file list
-            file_list.append((icon, i.name, i.is_dir(), size_f, units, str(timestamp_formatted)))
-
-
-    def get_file_list(self):
-        """retrieve self.file_list"""
-        return self.file_list
-
-    # callback signaled by Files_View
-    def cmp_f_list_dir_fst(self, model, row1, row2):
-        """
-        compare method for sorting sort columns in the file view
-        returns gt:1 lt:-1 or eq:0
-        """
-        sort_column, sort_order = model.get_sort_column_id()
-        name1 = model.get_value(row1, sort_column)
-        name2 = model.get_value(row2, sort_column)
-
-        if self.sort_ignore_case:
-            name1 = name1.lower()
-            name2 = name2.lower()
-
-        if self.sort_dir_first:
-            is_dir_1 = model.get_value(row1, 2)
-            is_dir_2 = model.get_value(row2, 2)
-            # account for the sort order when returning directories first
-            direction = 1
-            if sort_order is Gtk.SortType.DESCENDING:
-                direction = -1
-            #return immediately if comparing a dir and a file
-            if is_dir_1 and not is_dir_2:
-                return -1 * direction
-            if not is_dir_1 and is_dir_2:
-                return 1 * direction
-
-        if name1 < name2:
-            return -1
-        if name1 == name2:
-            return 0
-        return 1
-
-    def format_f_size(self, size):
-        """
-        convert filesize to string with appropriate units
-        This includes generating a units suffix thats returned with the formatted size as a tuple.
-        """
-        units = 'b'
-        length = len(f"{size:.0f}")
-        if length <= 3:
-            val = str(size)
-        elif length <= 6:
-            val = f"{size / 10e+2:.1f}"
-            units = 'kb'
-        elif length <= 9:
-            val = f"{size / 10e+5:.1f}"
-            units = 'mb'
-        elif length <= 12:
-            val = f"{size / 10e+8:.1f}"
-            units = 'gb'
-        else:
-            val = f"{size / 10e+11:.1f}"
-            units = 'tb'
-        return (val, units)
-
-    def append_to_path_back(self):
-        """track file change history"""
-        if len(self.path_back) >= self.path_back_max_len:
-            self.path_back.pop(0)
-        self.path_back.append(self.current_path)
-
-    def append_to_path_ahead(self):
-        """track file change history"""
-        if len(self.path_ahead) >= self.path_ahead_max_len:
-            self.path_ahead.pop(0)
-        self.path_ahead.append(self.current_path)
-
-
-    def get_path_current(self):
-        """get the current path"""
-        return self.current_path
-
-    def cd(self, path):
-        """move to a new working directory determined by path"""
-        if os.path.isdir(path):
-            self.append_to_path_back()
-            self.path_ahead.clear()
-            self.current_path = path
-            self.__update_file_list()
-
-    def cd_ahead(self):
-        """move forward to directory in the file change history"""
-        if len(self.path_ahead) > 0:
-            path = self.path_ahead.pop()
-            if os.path.isdir(path):
-                self.append_to_path_back()
-                self.current_path = path
-                self.__update_file_list()
-            else:
-                self.path_ahead.append(path)
-
-    def cd_up(self, path):
-        """move up one level in the directory tree"""
-        if os.path.isdir(path):
-            self.append_to_path_back()
-            self.cd(os.path.split(self.get_path_current())[0])
-            self.__update_file_list()
-
-    def cd_previous(self):
-        """move back to directory in the file change history"""
-        if len(self.path_back) > 0:
-            path = self.path_back.pop()
-            if os.path.isdir(path):
-                self.append_to_path_ahead()
-                self.current_path = path
-                self.__update_file_list()
-            else:
-                self.path_back.append(path)
-
-    def is_hidden_file(self, file_name):
-        """determine if a file is a hidden file"""
-        valid = re.compile(r"^[\.]")
-        if valid.match(file_name):
-            return True
-        return False
-
-
-class Files_View:
-    """Display file infrmation for files in the cwd"""
-
-    def __init__(self, files_view, files):
-        self.files_view = files_view
-        self.files_view.connect('destroy', self.on_destroy)
-        self.files_view_dbi = FilesViewDBI()
-        self.files = files
-        # set up the data model and containers
-        self.files_ls = self.files.get_file_list()
-        self.files_ls.set_sort_func(1, self.files.cmp_f_list_dir_fst, None)
-        self.files_view.set_model(self.files_ls)
-
-        # name column
-        name_r_icon = Gtk.CellRendererPixbuf()
-        name_r_text = Gtk.CellRendererText()
-        self.name_col = Gtk.TreeViewColumn("Name")
-        self.name_col.pack_start(name_r_icon, False)
-        self.name_col.pack_start(name_r_text, True)
-        self.name_col.add_attribute(name_r_icon, "pixbuf", 0)
-        self.name_col.add_attribute(name_r_text, "text", 1)
-        self.name_col.set_sort_column_id(1)
-        self.name_col.set_resizable(True)
-        # reset name column width to previous size iff previous size exists.
-        if name_width := self.files_view_dbi.get_name_col_width():
-            self.name_col.set_fixed_width(name_width)
-        self.files_view.append_column(self.name_col)
-
-        # size column
-        size_r_val = Gtk.CellRendererText()
-        size_r_units = Gtk.CellRendererText()
-        size_col = Gtk.TreeViewColumn("Size")
-        size_col.pack_start(size_r_val, False)
-        size_col.pack_start(size_r_units, False)
-        size_col.add_attribute(size_r_val, "text", 3)
-        size_col.add_attribute(size_r_units, "text", 4)
-        self.files_view.append_column(size_col)
-
-        # file creation time column
-        c_time_r = Gtk.CellRendererText()
-        c_time_col = Gtk.TreeViewColumn("Modified")
-        c_time_col.pack_start(c_time_r, True)
-        c_time_col.add_attribute(c_time_r, "text", 5)
-        self.files_view.append_column(c_time_col)
-
-        #signals
-        self.files_view.connect('row-activated', self.row_activated)
-        self.files_view.connect('button-release-event', self.on_button_release )
-
-    def on_button_release(self, unused_button, event):
-        """
-        handle mouse button release events.
-        Currently:
-        column resising
-        forward and back buttons to move ahead or back in the file change history.
-        """
-        if event.get_button()[0] is True:
-            if event.get_button()[1] == 1:
-                pass
-                #print('left button clicked')
-            elif event.get_button()[1] == 2:
-                pass
-                #print('middle button clicked')
-            elif event.get_button()[1] == 3:
-                pass
-                #print('right button clicked')
-            elif event.get_button()[1] == 8:
-                self.files.cd_previous()
-                #print('back button clicked')
-            elif event.get_button()[1] == 9:
-                self.files.cd_ahead()
-                #print('forward button clicked')
-
-    def row_activated(self, treeview, path, unused_column):
-        """
-        a file was cicked in the view.
-        if file is a directory then change to that directory
-        """
-        model = treeview.get_model()
-        tree_iter = model.get_iter(path)
-        value = model.get_value(tree_iter,1)
-        is_dir = model.get_value(tree_iter,2)
-        if is_dir:
-            # cd into selected dir
-            new_path = os.path.join(self.files.get_path_current(), value)
-            self.files.cd(new_path)
-
-    def on_destroy(self, *unused_args):
-        """save the gui's state"""
-        self.files_view_dbi.save_name_col_width(self.name_col.get_width())
-
-
-class FilesViewDBI:
-    """Class to help FilesView interface with a database"""
-
-    def __init__(self):
-        self.settings_numeric = book_ease_tables.SettingsNumeric()
-        # ids dict stores attribute:rowid to ease calls to update or insert a new row in the database
-        self.ids = {}
-
-    def get_name_col_width(self) -> int | None:
-        """retrieve the saved width of the name column in the FilesView treeview."""
-        with book_ease_tables.DB_CONNECTION_MANAGER.query() as con:
-            width_result = self.settings_numeric.get(con, 'FilesView', 'name_col_width')
-
-        if width_result:
-            width = width_result[0]['value']
-            self.ids['name_col_width'] = width_result[0]['id_']
-        else:
-            width = None
-            self.ids['name_col_width'] = None
-        return width
-
-    def save_name_col_width(self, width: int):
-        """Save the width of the name column in the FilesView:TreeView to a database."""
-        with book_ease_tables.DB_CONNECTION_MANAGER.query() as con:
-            if id_ := self.ids['name_col_width']:
-                self.settings_numeric.update_value_by_id(con, id_, width)
-            else:
-                self.settings_numeric.set(con, 'FilesView', 'name_col_width', width)
-
-
 class MainWindow(Gtk.Window):
     """The main display window"""
     SettingsNumericDBI = book_ease_tables.SettingsNumericDBI
@@ -906,9 +575,9 @@ def main(unused_args):  # pylint: disable=unused-variable
     builder = Gtk.Builder()
     builder.add_from_file("book_ease.glade")
     # files backend
-    files = Files_()
+    files = file_mgr.Files_()
     # left side file viewer
-    files_view_1 = Files_View(builder.get_object("files_1"), files)
+    files_view_1 = file_mgr_view.Files_View(builder.get_object("files_1"), files)
     # left side bookmarks
     book_mark_1 = BookMark(builder.get_object("bookmarks_1"), files_view_1, files)
     # image pane
